fix(KEN-1054): only an agreed, hash-shaped commit dates a report
