+++ changelog.d/changed/KEN-1054.md
@@ -1,1 +1,1 @@
-- `kendex report` stamps the reported asset's source commit and rendered hash from the lock into the issue marker, so triage can date a report against a fix.
+- `kendex report` stamps the source commit and rendered hash the lock recorded for the reported asset into the issue marker, so triage can date a report; installs that disagree stamp `unlocked`.
